A testing/development project in saving and retrieving data ("Load/Save")
At this point the project does not save to or read from a file, just serializes to and deserializes from a string.  The methodology involves the use of several custom iterators to parse the string and figure out what data goes where.

A 'string <class name>::Serialized()' method returns all saving data to a single line string, and a constructor of said class taking a single string ((serialString)) reconstructs an instance of the class with its data.